[refac/WEDPRO-183] refac: 이미지 슬라이더 컴포넌트로 분리
