+++ src/components/navbar/Navbar.jsx
@@ -175,8 +175,6 @@ function GuestNavbar({ setNavbar }) {
 }
 
 export default function Navbar({ setNavbar, token, guestState, uuid1 }) {
-  console.log(uuid1);
-  console.log(guestState);
   const [_, nickName] = useTokenDecode(token);
 
   const [navbarNotification, setNavbarNotification] = useState([]);
